ready to sumbit

+++ cpp09/ex02/PmergeMe.hpp
@@ -37,9 +37,9 @@ extern int	num_comparison;
 
 //PmergeMe main folder && printing
 void	PmergeMe_vector(int argc, char **argv);
-void	print_vector_pairs(std::vector<int>& vector, size_t pair_size, int mp);
-void	comparing_startVend(std::vector<int>& main, size_t start, size_t end, size_t pair_size);
-void	print_mainpend(std::vector<int> &main, std::vector<int> &pend, size_t pair_size, size_t before_after);
+void	print_vector_pairs(std::vector<long long>& vector, size_t pair_size, int mp);
+void	comparing_startVend(std::vector<long long>& main, size_t start, size_t end, size_t pair_size);
+void	print_mainpend(std::vector<long long> &main, std::vector<long long> &pend, size_t pair_size, size_t before_after);
 
 //--general
 long long   convert_token(std::string string);
@@ -48,16 +48,16 @@ size_t	jacob_number(size_t n);
 void	update_new_jacob(size_t &old_jacob, size_t &old_jacobsthal, size_t &n, size_t &jacobsthal);
 
 //--vector
-void	timer_for_recursion_vector(std::vector<int> &sorter, int argc);
-void	pushing_sort_vector(std::vector<int> &sorter, int argc, char **argv);
-void	print_vector_AFTER(std::vector<int> &vector, double timer_end, int argc);
-void	sortinside_pairs_vector(std::vector<int> &sorter, size_t power, int pair_counter);
-void	start_recursive_sort_vector(std::vector<int> &sorter, int level);
-int		counting_total_pair_vector(std::vector<int> &sorter, std::vector<int> &leftover, int power);
-std::vector<int>	sorting_mainpend_chain_vector(std::vector<int> &sorter, std::vector<int> &leftover, size_t power);
-void	insert_firstpair_vector(std::vector<int> &main, std::vector<int> &pend, size_t &pair_size);
-size_t	binary_search_leftover_vector(std::vector<int>& main, size_t start, size_t end, int insert_element, size_t pair_size);
-void	add_leftover_vector(std::vector<int> &main, std::vector<int> &leftover, size_t &pair_size);
+void	timer_for_recursion_vector(std::vector<long long> &sorter, int argc);
+void	pushing_sort_vector(std::vector<long long> &sorter, int argc, char **argv);
+void	print_vector_AFTER(std::vector<long long> &vector, double timer_end, int argc);
+void	sortinside_pairs_vector(std::vector<long long> &sorter, size_t power, int pair_counter);
+void	start_recursive_sort_vector(std::vector<long long> &sorter, int level);
+int		counting_total_pair_vector(std::vector<long long> &sorter, std::vector<long long> &leftover, int power);
+std::vector<long long>	sorting_mainpend_chain_vector(std::vector<long long> &sorter, std::vector<long long> &leftover, size_t power);
+void	insert_firstpair_vector(std::vector<long long> &main, std::vector<long long> &pend, size_t &pair_size);
+size_t	binary_search_leftover_vector(std::vector<long long>& main, size_t start, size_t end, int insert_element, size_t pair_size);
+void	add_leftover_vector(std::vector<long long> &main, std::vector<long long> &leftover, size_t &pair_size);
 
 //--deque
 void	PmergeMe_deque(int argc, char **argv);
